+++ lib/sanitize/config.rb
@@ -42,12 +42,6 @@ class Sanitize
       # that all HTML will be stripped).
       :elements => [],
 
-      # Output format. Supported formats are :html and :xhtml. Default is :html.
-      :output => :html,
-
-      # Character encoding to use for HTML output. Default is 'utf-8'.
-      :output_encoding => 'utf-8',
-
       # URL handling protocols to allow in specific attributes. By default, no
       # protocols are allowed. Use :relative in place of a protocol if you want
       # to allow relative URLs sans protocol.